Refactorisation mineure
