+++ src/front-end/typescript/lib/i18n/index.js
